Move Toolbar into separate component file
