+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/TitleBlockZen.scss
@@ -73,23 +73,6 @@
   }
 }
 
-.toolbar {
-  display: flex;
-  align-items: center;
-
-  .toolbarItem {
-    margin: 0 ($ca-grid / 4);
-  }
-
-  .toolbarItem:first-of-type {
-    @include ca-margin($start: 0, $end: $ca-grid / 4);
-  }
-
-  .toolbarItem:last-of-type {
-    @include ca-margin($start: $ca-grid / 4, $end: 0);
-  }
-}
-
 .tag {
   @include ca-margin($start: $ca-grid / 2);
   display: flex;
